+++ packages/controllers/src/controllers/BlockchainApiController.ts
@@ -298,15 +298,20 @@ export const BlockchainApiController = {
     })
   },
 
-  async getAddressBalance({ caipNetworkId, address }: BlockchainApiGetAddressBalanceRequest) {
+  async getAddressBalance<T = string>({
+    caipNetworkId,
+    address,
+    method = 'getAddressBalance',
+    params
+  }: BlockchainApiGetAddressBalanceRequest) {
     return state.api
-      .post<BlockchainApiGetAddressBalanceResponse>({
+      .post<BlockchainApiGetAddressBalanceResponse<T>>({
         path: `/v1?chainId=${caipNetworkId}&projectId=${OptionsController.state.projectId}`,
         body: {
           id: '1',
           jsonrpc: '2.0',
-          method: 'getAddressBalance',
-          params: { address }
+          method,
+          params: params ?? { address }
         }
       })
       .then(result => result.result)