+++ packages/controllers/src/utils/AssetUtil.ts
@@ -32,7 +32,7 @@ const namespaceImageIds: Record<ChainNamespace, string> = {
   // TON
   ton: '20f673c0-095e-49b2-07cf-eb5049dcf600',
   // TRON
-  tron: 'dd9de794-d4ce-4c94-682f-a367f926d500'
+  tron: '3502bb86-cc4e-420f-a387-59ea63a28b00'
 }
 
 // -- State --------------------------------------------- //