+++ packages/ui/src/utils/TypeUtil.ts
@@ -184,6 +184,7 @@ export type IconType =
   | 'telegram'
   | 'threeDots'
   | 'ton'
+  | 'tron'
   | 'twitch'
   | 'twitter'
   | 'twitterIcon'