+++ packages/siwe/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @reown/appkit-siwe
 
+## 1.8.19
+
+### Patch Changes
+
+- [#5526](https://github.com/reown-com/appkit/pull/5526) [`a7646ab`](https://github.com/reown-com/appkit/commit/a7646ab7f3aacb62a8ef23e9670cf41a77609730) Thanks [@tomiir](https://github.com/tomiir)! - Adds more options to fetchWallets method on useAppKitWallets
+
+- [#5535](https://github.com/reown-com/appkit/pull/5535) [`249beb6`](https://github.com/reown-com/appkit/commit/249beb64220596d468c0f91de7b568f5f640b26f) Thanks [@tomiir](https://github.com/tomiir)! - Serializes caip address and chain Id into allAccounts
+
+- Updated dependencies [[`a7646ab`](https://github.com/reown-com/appkit/commit/a7646ab7f3aacb62a8ef23e9670cf41a77609730), [`6851c1b`](https://github.com/reown-com/appkit/commit/6851c1ba88c183bfc5b6ca1bf8e0efba26012b78), [`249beb6`](https://github.com/reown-com/appkit/commit/249beb64220596d468c0f91de7b568f5f640b26f)]:
+  - @reown/appkit-utils@1.8.19
+  - @reown/appkit-controllers@1.8.19
+  - @reown/appkit-common@1.8.19
+  - @reown/appkit-wallet@1.8.19
+  - @reown/appkit-ui@1.8.19
+
 ## 1.8.18
 
 ### Patch Changes